refactor: move levels menu functions to own module manage_menu.js
