+++ nonograms/scripts/game_logic.js
@@ -11,42 +11,6 @@ import { Popup } from './ui_components/popup.js';
 let isOverlayShown = false;
 let currentCell = null;
 let prevCell = null;
-let prevLevel = null;
-let prevPicture = null;
-let menuLevel = null;
-
-function tapOnMenu(event) {
-  const isMenu = event.target.classList.contains('level__menu_item');
-  const isMenuLevel = event.target.classList.contains('level__item');
-  const isLevelImage = event.target.classList.contains('level__image');
-
-  if (isMenu || isMenuLevel || isLevelImage) {
-    event.preventDefault();
-  }
-
-  if (isMenu) {
-    menuLevel = event.target;
-    event.target.classList.toggle('menu__tap');
-    isOverlayShown ? removeOverlay(event) : setOverlay(event);
-    return;
-  }
-
-  if (isMenuLevel && event.target !== prevLevel) {
-    event.target.classList.add('menu__level_hover');
-    prevLevel?.classList.remove('menu__level_hover');
-    prevPicture?.classList.remove('menu__image_hover');
-    prevLevel = event.target;
-    prevPicture = null;
-    return;
-  }
-
-  if (isLevelImage && event.target !== prevPicture) {
-    event.target.classList.add('menu__image_hover');
-    prevPicture?.classList.remove('menu__image_hover');
-    prevPicture = event.target;
-    manageHeader(event);
-  }
-}
 
 function mouseButtonUp() {
   currentCell = null;
@@ -136,16 +100,6 @@ function removeOverlay(event) {
   isOverlayShown = false;
   listeners.setOverlayListeners(event, false);
   resetMenus();
-}
-
-function resetMenus() {
-  const sidebar = document.getElementsByClassName('sidebar');
-  sidebar[0]?.classList.remove('sidebar__active');
-  menuLevel?.classList.remove('menu__tap');
-  prevLevel?.classList.remove('menu__level_hover');
-  prevPicture?.classList.remove('menu__image_hover');
-  prevLevel = null;
-  prevPicture = null;
 }
 
 function toggleSidebar() {
